requirements.txt
Flask==3.1.2
langchain==0.3.27
langchain_community==0.3.30
langchain_core==0.3.76
langchain_openai==0.3.33
pydantic==2.11.9
python-dotenv==1.1.1
tweepy==4.16.0